Original channel changer that analyzes the screen
When these STBs go to sleep, it activates a screensaver, and you have to
press "Ok" to wake it up before pressing the number buttons to change
the channel.

This script takes a short video clip and uses jp2a to convert it to
a series of text files with color information as escape codes.  It tries
to detect whether it's in screensaver mode, or showing live video.  It's
more complicated because the screensaver isn't a static image, it's a mostly
black screen with moving, blue "smoke".  The words "Press OK to watch ATT
Uverse" are in white and move around the screen.